+++ test/test_sentence.py
@@ -77,7 +77,7 @@ def test_cg_conv_equivalence():
     p1 = Popen(f'hfst-lookup {RSRC_PATH}analyser-gt-desc.hfstol | cg-conv -fC',  # noqa: E501
                stdin=PIPE, stdout=PIPE, universal_newlines=True, shell=True)
     output, error = p1.communicate(toks)
-    assert output == sent.cg3_str()
+    assert output == sent.cg3_str(annotated=False) + '\n'
 
 
 def test_cg3_parse():
@@ -89,7 +89,7 @@ def test_cg3_parse():
     p1 = Popen(f'hfst-lookup {RSRC_PATH}analyser-gt-desc.hfstol | cg-conv -fC | vislcg3 -g {RSRC_PATH}disambiguator.cg3',  # noqa: E501
                stdin=PIPE, stdout=PIPE, universal_newlines=True, shell=True)
     output, error = p1.communicate(toks)
-    assert output == sent.cg3_str()
+    assert output == sent.cg3_str(annotated=False) + '\n'
 
 
 def test_cg3_parse_w_traces():
@@ -101,7 +101,7 @@ def test_cg3_parse_w_traces():
     p1 = Popen(f'hfst-lookup {RSRC_PATH}analyser-gt-desc.hfstol | cg-conv -fC | vislcg3 -t -g {RSRC_PATH}disambiguator.cg3',  # noqa: E501
                stdin=PIPE, stdout=PIPE, universal_newlines=True, shell=True)
     output, error = p1.communicate(toks)
-    assert output == sent.cg3_str(traces=True)
+    assert output == sent.cg3_str(annotated=False, traces=True) + '\n'
 
 
 def test_from_hfst():
