+++ udar/__init__.py
@@ -5,9 +5,10 @@ from .reading import *  # noqa: F401, F403
 from .tok import *  # noqa: F401, F403
 from .sentence import *  # noqa: F401, F403
 from .document import *  # noqa: F401, F403
-from .features import *  # noqa: F401, F403
 from .fsts import *  # noqa: F401, F403
 
 from .convenience import *  # noqa: F401, F403
 from .misc import *  # noqa: F401, F403
 from .transliterate import *  # noqa: F401, F403
+
+from .features import *  # noqa: F401, F403
